refactor: re-design login flow
