+++ src/containers/auth/Layout.tsx
@@ -1,12 +1,12 @@
 import Image from "next/image";
 import Footer from "@/components/Footer";
 
-const HomeLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => (
+const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => (
   <section className="outer">
     <div className="inner flex flex-col h-screen">
       <section className="flex-grow grid grid-cols-12 gap-8">
         {/* 圖片 */}
-        <section className="col-span-6 flex flex-col justify-center gap-12 pl-12">
+        <header className="col-span-6 flex flex-col justify-center gap-12 pl-12">
           <div className="flex items-center gap-4">
             <Image
               src="/images/logo-rounded.svg"
@@ -27,13 +27,17 @@ const HomeLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => (
             width={563}
             height={70}
           />
-        </section>
+        </header>
         {/* 表單*/}
-        {children}
+        <div className="col-span-5 col-start-8 flex flex-col justify-center pr-12">
+          <section className="flex flex-col justify-center gap-8 p-8 border border-stroke rounded-[30px]">
+            {children}
+          </section>
+        </div>
       </section>
     </div>
     <Footer />
   </section>
 );
 
-export default HomeLayout;
+export default Layout;
